+++ src/types/tournament.ts
@@ -1,0 +1,12 @@
+export type Period = {
+  ID: number;
+  PeriodType: "break" | "round";
+  RoundID: null | number;
+  StartTime: string;
+  EndTime: string; // iso date
+};
+
+export type PeriodsResponse = {
+  allPeriods: Period[];
+  activePeriod: Period;
+};